Network.AWS.MechanicalTurk

Description

Amazon Mechanical Turk API Reference

Service Configuration

mechanicalTurk :: Service

API version 2017-01-17 of the Amazon Mechanical Turk SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MechanicalTurk.

ServiceFault

_ServiceFault :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Mechanical Turk is temporarily unable to process your request. Try your call again.

RequestError

_RequestError :: AsError a => Getting (First ServiceError) a ServiceError

Your request is invalid.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ApproveAssignment

module Network.AWS.MechanicalTurk.ApproveAssignment

ListReviewPolicyResultsForHIT

module Network.AWS.MechanicalTurk.ListReviewPolicyResultsForHIT

ListHITs

module Network.AWS.MechanicalTurk.ListHITs

ListWorkersWithQualificationType

module Network.AWS.MechanicalTurk.ListWorkersWithQualificationType

DeleteHIT

module Network.AWS.MechanicalTurk.DeleteHIT

ListReviewableHITs

module Network.AWS.MechanicalTurk.ListReviewableHITs

GetAssignment

module Network.AWS.MechanicalTurk.GetAssignment

DeleteQualificationType

module Network.AWS.MechanicalTurk.DeleteQualificationType

UpdateQualificationType

module Network.AWS.MechanicalTurk.UpdateQualificationType

ListQualificationTypes

module Network.AWS.MechanicalTurk.ListQualificationTypes

UpdateHITTypeOfHIT

module Network.AWS.MechanicalTurk.UpdateHITTypeOfHIT

DisassociateQualificationFromWorker

module Network.AWS.MechanicalTurk.DisassociateQualificationFromWorker

SendTestEventNotification

module Network.AWS.MechanicalTurk.SendTestEventNotification

NotifyWorkers

module Network.AWS.MechanicalTurk.NotifyWorkers

CreateHITWithHITType

module Network.AWS.MechanicalTurk.CreateHITWithHITType

CreateHITType

module Network.AWS.MechanicalTurk.CreateHITType

SendBonus

module Network.AWS.MechanicalTurk.SendBonus

ListQualificationRequests

module Network.AWS.MechanicalTurk.ListQualificationRequests

UpdateExpirationForHIT

module Network.AWS.MechanicalTurk.UpdateExpirationForHIT

RejectAssignment

module Network.AWS.MechanicalTurk.RejectAssignment

ListAssignmentsForHIT

module Network.AWS.MechanicalTurk.ListAssignmentsForHIT

RejectQualificationRequest

module Network.AWS.MechanicalTurk.RejectQualificationRequest

GetQualificationScore

module Network.AWS.MechanicalTurk.GetQualificationScore

GetQualificationType

module Network.AWS.MechanicalTurk.GetQualificationType

UpdateHITReviewStatus

module Network.AWS.MechanicalTurk.UpdateHITReviewStatus

CreateQualificationType

module Network.AWS.MechanicalTurk.CreateQualificationType

AcceptQualificationRequest

module Network.AWS.MechanicalTurk.AcceptQualificationRequest

GetFileUploadURL

module Network.AWS.MechanicalTurk.GetFileUploadURL

CreateAdditionalAssignmentsForHIT

module Network.AWS.MechanicalTurk.CreateAdditionalAssignmentsForHIT

GetHIT

module Network.AWS.MechanicalTurk.GetHIT

CreateWorkerBlock

module Network.AWS.MechanicalTurk.CreateWorkerBlock

ListHITsForQualificationType

module Network.AWS.MechanicalTurk.ListHITsForQualificationType

ListBonusPayments

module Network.AWS.MechanicalTurk.ListBonusPayments

ListWorkerBlocks

module Network.AWS.MechanicalTurk.ListWorkerBlocks

DeleteWorkerBlock

module Network.AWS.MechanicalTurk.DeleteWorkerBlock

UpdateNotificationSettings

module Network.AWS.MechanicalTurk.UpdateNotificationSettings

AssociateQualificationWithWorker

module Network.AWS.MechanicalTurk.AssociateQualificationWithWorker

CreateHIT

module Network.AWS.MechanicalTurk.CreateHIT

GetAccountBalance

module Network.AWS.MechanicalTurk.GetAccountBalance

Types

AssignmentStatus

data AssignmentStatus

Constructors

Approved

Rejected

Submitted

Comparator

data Comparator

Constructors

DoesNotExist

EqualTo

Exists

GreaterThan

GreaterThanOrEqualTo

IN

LessThan

LessThanOrEqualTo

NotEqualTo

NotIn

EventType

data EventType

Constructors

AssignmentAbandoned

AssignmentAccepted

AssignmentApproved

AssignmentRejected

AssignmentReturned

AssignmentSubmitted

HITCreated

HITDisposed

HITExpired

HITExtended

HITReviewable

Ping

HITReviewStatus

data HITReviewStatus

Constructors

MarkedForReview

NotReviewed

ReviewedAppropriate

ReviewedInappropriate

HITStatus

data HITStatus

Constructors

HITSAssignable

HITSDisposed

HITSReviewable

HITSReviewing

HITSUnassignable

NotificationTransport

data NotificationTransport

Constructors

Email

Sqs

NotifyWorkersFailureCode

data NotifyWorkersFailureCode

Constructors

HardFailure

SoftFailure

QualificationStatus

data QualificationStatus

Constructors

Granted

Revoked

QualificationTypeStatus

data QualificationTypeStatus

Constructors

Active

Inactive

ReviewActionStatus

data ReviewActionStatus

Constructors

Cancelled

Failed

Intended

Succeeded

ReviewPolicyLevel

data ReviewPolicyLevel

Constructors

Assignment

Hit

ReviewableHITStatus

data ReviewableHITStatus

Constructors

Reviewable

Reviewing

Assignment

data Assignment

The Assignment data structure represents a single assignment of a HIT to a Worker. The assignment tracks the Worker's efforts to complete the HIT, and contains the results for later retrieval.

See: assignment smart constructor.

assignment :: Assignment

Creates a value of Assignment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aAcceptTime - The date and time the Worker accepted the assignment.

aAnswer - The Worker's answers submitted for the HIT contained in a QuestionFormAnswers document, if the Worker provides an answer. If the Worker does not provide any answers, Answer may contain a QuestionFormAnswers document, or Answer may be empty.

aAssignmentStatus - The status of the assignment.

aRequesterFeedback - The feedback string included with the call to the ApproveAssignment operation or the RejectAssignment operation, if the Requester approved or rejected the assignment and specified feedback.

aDeadline - The date and time of the deadline for the assignment. This value is derived from the deadline specification for the HIT and the date and time the Worker accepted the HIT.

aApprovalTime - If the Worker has submitted results and the Requester has approved the results, ApprovalTime is the date and time the Requester approved the results. This value is omitted from the assignment if the Requester has not yet approved the results.

aRejectionTime - If the Worker has submitted results and the Requester has rejected the results, RejectionTime is the date and time the Requester rejected the results.

aAutoApprovalTime - If results have been submitted, AutoApprovalTime is the date and time the results of the assignment results are considered Approved automatically if they have not already been explicitly approved or rejected by the Requester. This value is derived from the auto-approval delay specified by the Requester in the HIT. This value is omitted from the assignment if the Worker has not yet submitted results.

aHITId - The ID of the HIT.

aWorkerId - The ID of the Worker who accepted the HIT.

aAssignmentId - A unique identifier for the assignment.

aSubmitTime - If the Worker has submitted results, SubmitTime is the date and time the assignment was submitted. This value is omitted from the assignment if the Worker has not yet submitted results.

aAcceptTime :: Lens' Assignment (Maybe UTCTime)

The date and time the Worker accepted the assignment.

aAnswer :: Lens' Assignment (Maybe Text)

The Worker's answers submitted for the HIT contained in a QuestionFormAnswers document, if the Worker provides an answer. If the Worker does not provide any answers, Answer may contain a QuestionFormAnswers document, or Answer may be empty.

aAssignmentStatus :: Lens' Assignment (Maybe AssignmentStatus)

The status of the assignment.

aRequesterFeedback :: Lens' Assignment (Maybe Text)

The feedback string included with the call to the ApproveAssignment operation or the RejectAssignment operation, if the Requester approved or rejected the assignment and specified feedback.

aDeadline :: Lens' Assignment (Maybe UTCTime)

The date and time of the deadline for the assignment. This value is derived from the deadline specification for the HIT and the date and time the Worker accepted the HIT.

aApprovalTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results and the Requester has approved the results, ApprovalTime is the date and time the Requester approved the results. This value is omitted from the assignment if the Requester has not yet approved the results.

aRejectionTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results and the Requester has rejected the results, RejectionTime is the date and time the Requester rejected the results.

aAutoApprovalTime :: Lens' Assignment (Maybe UTCTime)

If results have been submitted, AutoApprovalTime is the date and time the results of the assignment results are considered Approved automatically if they have not already been explicitly approved or rejected by the Requester. This value is derived from the auto-approval delay specified by the Requester in the HIT. This value is omitted from the assignment if the Worker has not yet submitted results.

aHITId :: Lens' Assignment (Maybe Text)

The ID of the HIT.

aWorkerId :: Lens' Assignment (Maybe Text)

The ID of the Worker who accepted the HIT.

aAssignmentId :: Lens' Assignment (Maybe Text)

A unique identifier for the assignment.

aSubmitTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results, SubmitTime is the date and time the assignment was submitted. This value is omitted from the assignment if the Worker has not yet submitted results.

BonusPayment

data BonusPayment

An object representing a Bonus payment paid to a Worker.

See: bonusPayment smart constructor.

bonusPayment :: BonusPayment

Creates a value of BonusPayment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bpReason - The Reason text given when the bonus was granted, if any.

bpGrantTime - The date and time of when the bonus was granted.

bpWorkerId - The ID of the Worker to whom the bonus was paid.

bpAssignmentId - The ID of the assignment associated with this bonus payment.

bpBonusAmount - Undocumented member.

bpReason :: Lens' BonusPayment (Maybe Text)

The Reason text given when the bonus was granted, if any.

bpGrantTime :: Lens' BonusPayment (Maybe UTCTime)

The date and time of when the bonus was granted.

bpWorkerId :: Lens' BonusPayment (Maybe Text)

The ID of the Worker to whom the bonus was paid.

bpAssignmentId :: Lens' BonusPayment (Maybe Text)

The ID of the assignment associated with this bonus payment.

bpBonusAmount :: Lens' BonusPayment (Maybe Text)

Undocumented member.

HIT

data HIT

The HIT data structure represents a single HIT, including all the information necessary for a Worker to accept and complete the HIT.

See: hIT smart constructor.

hIT :: HIT

Creates a value of HIT with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hitCreationTime - The date and time the HIT was created.

hitHITGroupId - The ID of the HIT Group of this HIT.

hitNumberOfAssignmentsPending - The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

hitHITTypeId - The ID of the HIT type of this HIT

hitExpiration - The date and time the HIT expires.

hitAutoApprovalDelayInSeconds - The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

hitRequesterAnnotation - An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

hitHITStatus - The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

hitMaxAssignments - The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

hitNumberOfAssignmentsCompleted - The number of assignments for this HIT that have been approved or rejected.

hitReward - Undocumented member.

hitKeywords - One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.

hitHITLayoutId - The ID of the HIT Layout of this HIT.

hitQualificationRequirements - A condition that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met by a Worker's Qualifications for the Worker to accept the HIT.

hitTitle - The title of the HIT.

hitHITId - A unique identifier for the HIT.

hitHITReviewStatus - Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

hitNumberOfAssignmentsAvailable - The number of assignments for this HIT that are available for Workers to accept.

hitDescription - A general description of the HIT.

hitQuestion - The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

hitAssignmentDurationInSeconds - The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

hitCreationTime :: Lens' HIT (Maybe UTCTime)

The date and time the HIT was created.

hitHITGroupId :: Lens' HIT (Maybe Text)

The ID of the HIT Group of this HIT.

hitNumberOfAssignmentsPending :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

hitHITTypeId :: Lens' HIT (Maybe Text)

The ID of the HIT type of this HIT

hitExpiration :: Lens' HIT (Maybe UTCTime)

The date and time the HIT expires.

hitAutoApprovalDelayInSeconds :: Lens' HIT (Maybe Integer)

The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

hitRequesterAnnotation :: Lens' HIT (Maybe Text)

An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

hitHITStatus :: Lens' HIT (Maybe HITStatus)

The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

hitMaxAssignments :: Lens' HIT (Maybe Int)

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

hitNumberOfAssignmentsCompleted :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that have been approved or rejected.

hitReward :: Lens' HIT (Maybe Text)

Undocumented member.

hitKeywords :: Lens' HIT (Maybe Text)

One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.

hitHITLayoutId :: Lens' HIT (Maybe Text)

The ID of the HIT Layout of this HIT.

hitQualificationRequirements :: Lens' HIT [QualificationRequirement]

A condition that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met by a Worker's Qualifications for the Worker to accept the HIT.

hitTitle :: Lens' HIT (Maybe Text)

The title of the HIT.

hitHITId :: Lens' HIT (Maybe Text)

A unique identifier for the HIT.

hitHITReviewStatus :: Lens' HIT (Maybe HITReviewStatus)

Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

hitNumberOfAssignmentsAvailable :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that are available for Workers to accept.

hitDescription :: Lens' HIT (Maybe Text)

A general description of the HIT.

hitQuestion :: Lens' HIT (Maybe Text)

The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

hitAssignmentDurationInSeconds :: Lens' HIT (Maybe Integer)

The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

HITLayoutParameter

data HITLayoutParameter

The HITLayoutParameter data structure defines parameter values used with a HITLayout. A HITLayout is a reusable Amazon Mechanical Turk project template used to provide Human Intelligence Task (HIT) question data for CreateHIT.

See: hITLayoutParameter smart constructor.

hITLayoutParameter :: HITLayoutParameter

Creates a value of HITLayoutParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hitlpValue - The value substituted for the parameter referenced in the HITLayout.

hitlpName - The name of the parameter in the HITLayout.

hitlpValue :: Lens' HITLayoutParameter (Maybe Text)

The value substituted for the parameter referenced in the HITLayout.

hitlpName :: Lens' HITLayoutParameter (Maybe Text)

The name of the parameter in the HITLayout.

Locale

data Locale

The Locale data structure represents a geographical region or location.

See: locale smart constructor.

locale

Arguments

:: Text
lCountry

-> Locale

Creates a value of Locale with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lSubdivision - The state or subdivision of the locale. A valid ISO 3166-2 subdivision code. For example, the code WA refers to the state of Washington.

lCountry - The country of the locale. Must be a valid ISO 3166 country code. For example, the code US refers to the United States of America.

lSubdivision :: Lens' Locale (Maybe Text)

The state or subdivision of the locale. A valid ISO 3166-2 subdivision code. For example, the code WA refers to the state of Washington.

lCountry :: Lens' Locale Text

The country of the locale. Must be a valid ISO 3166 country code. For example, the code US refers to the United States of America.

NotificationSpecification

data NotificationSpecification

The NotificationSpecification data structure describes a HIT event notification for a HIT type.

See: notificationSpecification smart constructor.

notificationSpecification

Arguments

:: Text
nsDestination

-> NotificationTransport
nsTransport

-> NotificationSpecification

Creates a value of NotificationSpecification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nsEventTypes - The list of events that should cause notifications to be sent. Valid Values: AssignmentAccepted | AssignmentAbandoned | AssignmentReturned | AssignmentSubmitted | AssignmentRejected | AssignmentApproved | HITCreated | HITExtended | HITDisposed | HITReviewable | HITExpired | Ping. The Ping event is only valid for the SendTestEventNotification operation.

nsVersion - The version of the Notification API to use. Valid value is 2006-05-05.

nsDestination - The destination for notification messages. or email notifications (if Transport is Email), this is an email address. For Amazon Simple Queue Service (Amazon SQS) notifications (if Transport is SQS), this is the URL for your Amazon SQS queue.

nsTransport - The method Amazon Mechanical Turk uses to send the notification. Valid Values: Email | SQS.

nsEventTypes :: Lens' NotificationSpecification [EventType]

The list of events that should cause notifications to be sent. Valid Values: AssignmentAccepted | AssignmentAbandoned | AssignmentReturned | AssignmentSubmitted | AssignmentRejected | AssignmentApproved | HITCreated | HITExtended | HITDisposed | HITReviewable | HITExpired | Ping. The Ping event is only valid for the SendTestEventNotification operation.

nsVersion :: Lens' NotificationSpecification (Maybe Text)

The version of the Notification API to use. Valid value is 2006-05-05.

nsDestination :: Lens' NotificationSpecification Text

The destination for notification messages. or email notifications (if Transport is Email), this is an email address. For Amazon Simple Queue Service (Amazon SQS) notifications (if Transport is SQS), this is the URL for your Amazon SQS queue.

nsTransport :: Lens' NotificationSpecification NotificationTransport

The method Amazon Mechanical Turk uses to send the notification. Valid Values: Email | SQS.

NotifyWorkersFailureStatus

data NotifyWorkersFailureStatus

When MTurk encounters an issue with notifying the Workers you specified, it returns back this object with failure details.

See: notifyWorkersFailureStatus smart constructor.

notifyWorkersFailureStatus :: NotifyWorkersFailureStatus

Creates a value of NotifyWorkersFailureStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nwfsNotifyWorkersFailureMessage - A message detailing the reason the Worker could not be notified.

nwfsNotifyWorkersFailureCode - Encoded value for the failure type.

nwfsWorkerId - The ID of the Worker.

nwfsNotifyWorkersFailureMessage :: Lens' NotifyWorkersFailureStatus (Maybe Text)

A message detailing the reason the Worker could not be notified.

nwfsNotifyWorkersFailureCode :: Lens' NotifyWorkersFailureStatus (Maybe NotifyWorkersFailureCode)

Encoded value for the failure type.

nwfsWorkerId :: Lens' NotifyWorkersFailureStatus (Maybe Text)

The ID of the Worker.

ParameterMapEntry

data ParameterMapEntry

This data structure is the data type for the AnswerKey parameter of the ScoreMyKnownAnswers/2011-09-01 Review Policy.

See: parameterMapEntry smart constructor.

parameterMapEntry :: ParameterMapEntry

Creates a value of ParameterMapEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pmeValues - The list of answers to the question specified in the MapEntry Key element. The Worker must match all values in order for the answer to be scored correctly.

pmeKey - The QuestionID from the HIT that is used to identify which question requires Mechanical Turk to score as part of the ScoreMyKnownAnswers/2011-09-01 Review Policy.

pmeValues :: Lens' ParameterMapEntry [Text]

The list of answers to the question specified in the MapEntry Key element. The Worker must match all values in order for the answer to be scored correctly.

pmeKey :: Lens' ParameterMapEntry (Maybe Text)

The QuestionID from the HIT that is used to identify which question requires Mechanical Turk to score as part of the ScoreMyKnownAnswers/2011-09-01 Review Policy.

PolicyParameter

data PolicyParameter

Name of the parameter from the Review policy.

See: policyParameter smart constructor.

policyParameter :: PolicyParameter

Creates a value of PolicyParameter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ppValues - The list of values of the Parameter

ppMapEntries - List of ParameterMapEntry objects.

ppKey - Name of the parameter from the list of Review Polices.

ppValues :: Lens' PolicyParameter [Text]

The list of values of the Parameter

ppMapEntries :: Lens' PolicyParameter [ParameterMapEntry]

List of ParameterMapEntry objects.

ppKey :: Lens' PolicyParameter (Maybe Text)

Name of the parameter from the list of Review Polices.

Qualification

data Qualification

The Qualification data structure represents a Qualification assigned to a user, including the Qualification type and the value (score).

See: qualification smart constructor.

qualification :: Qualification

Creates a value of Qualification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qStatus - The status of the Qualification. Valid values are Granted | Revoked.

qIntegerValue - The value (score) of the Qualification, if the Qualification has an integer value.

qLocaleValue - Undocumented member.

qQualificationTypeId - The ID of the Qualification type for the Qualification.

qGrantTime - The date and time the Qualification was granted to the Worker. If the Worker's Qualification was revoked, and then re-granted based on a new Qualification request, GrantTime is the date and time of the last call to the AcceptQualificationRequest operation.

qWorkerId - The ID of the Worker who possesses the Qualification.

qStatus :: Lens' Qualification (Maybe QualificationStatus)

The status of the Qualification. Valid values are Granted | Revoked.

qIntegerValue :: Lens' Qualification (Maybe Int)

The value (score) of the Qualification, if the Qualification has an integer value.

qLocaleValue :: Lens' Qualification (Maybe Locale)

Undocumented member.

qQualificationTypeId :: Lens' Qualification (Maybe Text)

The ID of the Qualification type for the Qualification.

qGrantTime :: Lens' Qualification (Maybe UTCTime)

The date and time the Qualification was granted to the Worker. If the Worker's Qualification was revoked, and then re-granted based on a new Qualification request, GrantTime is the date and time of the last call to the AcceptQualificationRequest operation.

qWorkerId :: Lens' Qualification (Maybe Text)

The ID of the Worker who possesses the Qualification.

QualificationRequest

data QualificationRequest

The QualificationRequest data structure represents a request a Worker has made for a Qualification.

See: qualificationRequest smart constructor.

qualificationRequest :: QualificationRequest

Creates a value of QualificationRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

quaQualificationRequestId - The ID of the Qualification request, a unique identifier generated when the request was submitted.

quaTest - The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

quaQualificationTypeId - The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

quaAnswer - The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

quaWorkerId - The ID of the Worker requesting the Qualification.

quaSubmitTime - The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.

quaQualificationRequestId :: Lens' QualificationRequest (Maybe Text)

The ID of the Qualification request, a unique identifier generated when the request was submitted.

quaTest :: Lens' QualificationRequest (Maybe Text)

The contents of the Qualification test that was presented to the Worker, if the type has a test and the Worker has submitted answers. This value is identical to the QuestionForm associated with the Qualification type at the time the Worker requests the Qualification.

quaQualificationTypeId :: Lens' QualificationRequest (Maybe Text)

The ID of the Qualification type the Worker is requesting, as returned by the CreateQualificationType operation.

quaAnswer :: Lens' QualificationRequest (Maybe Text)

The Worker's answers for the Qualification type's test contained in a QuestionFormAnswers document, if the type has a test and the Worker has submitted answers. If the Worker does not provide any answers, Answer may be empty.

quaWorkerId :: Lens' QualificationRequest (Maybe Text)

The ID of the Worker requesting the Qualification.

quaSubmitTime :: Lens' QualificationRequest (Maybe UTCTime)

The date and time the Qualification request had a status of Submitted. This is either the time the Worker submitted answers for a Qualification test, or the time the Worker requested the Qualification if the Qualification type does not have a test.

QualificationRequirement

data QualificationRequirement

The QualificationRequirement data structure describes a Qualification that a Worker must have before the Worker is allowed to accept a HIT. A requirement may optionally state that a Worker must have the Qualification in order to preview the HIT.

See: qualificationRequirement smart constructor.

qualificationRequirement

Arguments

:: Text
qrQualificationTypeId

-> Comparator
qrComparator

-> QualificationRequirement

Creates a value of QualificationRequirement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qrLocaleValues - The locale value to compare against the Qualification's value. The local value must be a valid ISO 3166 country code or supports ISO 3166-2 subdivisions. LocaleValue can only be used with a Worker_Locale QualificationType ID. LocaleValue can only be used with the EqualTo, NotEqualTo, In, and NotIn comparators. You must only use a single LocaleValue element when using the EqualTo or NotEqualTo comparators. When performing a set comparison by using the In or the NotIn comparator, you can use up to 30 LocaleValue elements in a QualificationRequirement data structure.

qrRequiredToPreview - If true, the question data for the HIT will not be shown when a Worker whose Qualifications do not meet this requirement tries to preview the HIT. That is, a Worker's Qualifications must meet all of the requirements for which RequiredToPreview is true in order to preview the HIT. If a Worker meets all of the requirements where RequiredToPreview is true (or if there are no such requirements), but does not meet all of the requirements for the HIT, the Worker will be allowed to preview the HIT's question data, but will not be allowed to accept and complete the HIT. The default is false.

qrIntegerValues - The integer value to compare against the Qualification's value. IntegerValue must not be present if Comparator is Exists or DoesNotExist. IntegerValue can only be used if the Qualification type has an integer value; it cannot be used with the Worker_Locale QualificationType ID. When performing a set comparison by using the In or the NotIn comparator, you can use up to 15 IntegerValue elements in a QualificationRequirement data structure.

qrQualificationTypeId - The ID of the Qualification type for the requirement.

qrComparator - The kind of comparison to make against a Qualification's value. You can compare a Qualification's value to an IntegerValue to see if it is LessThan, LessThanOrEqualTo, GreaterThan, GreaterThanOrEqualTo, EqualTo, or NotEqualTo the IntegerValue. You can compare it to a LocaleValue to see if it is EqualTo, or NotEqualTo the LocaleValue. You can check to see if the value is In or NotIn a set of IntegerValue or LocaleValue values. Lastly, a Qualification requirement can also test if a Qualification Exists or DoesNotExist in the user's profile, regardless of its value.

qrLocaleValues :: Lens' QualificationRequirement [Locale]

The locale value to compare against the Qualification's value. The local value must be a valid ISO 3166 country code or supports ISO 3166-2 subdivisions. LocaleValue can only be used with a Worker_Locale QualificationType ID. LocaleValue can only be used with the EqualTo, NotEqualTo, In, and NotIn comparators. You must only use a single LocaleValue element when using the EqualTo or NotEqualTo comparators. When performing a set comparison by using the In or the NotIn comparator, you can use up to 30 LocaleValue elements in a QualificationRequirement data structure.

qrRequiredToPreview :: Lens' QualificationRequirement (Maybe Bool)

If true, the question data for the HIT will not be shown when a Worker whose Qualifications do not meet this requirement tries to preview the HIT. That is, a Worker's Qualifications must meet all of the requirements for which RequiredToPreview is true in order to preview the HIT. If a Worker meets all of the requirements where RequiredToPreview is true (or if there are no such requirements), but does not meet all of the requirements for the HIT, the Worker will be allowed to preview the HIT's question data, but will not be allowed to accept and complete the HIT. The default is false.

qrIntegerValues :: Lens' QualificationRequirement [Int]

The integer value to compare against the Qualification's value. IntegerValue must not be present if Comparator is Exists or DoesNotExist. IntegerValue can only be used if the Qualification type has an integer value; it cannot be used with the Worker_Locale QualificationType ID. When performing a set comparison by using the In or the NotIn comparator, you can use up to 15 IntegerValue elements in a QualificationRequirement data structure.

qrQualificationTypeId :: Lens' QualificationRequirement Text

The ID of the Qualification type for the requirement.

qrComparator :: Lens' QualificationRequirement Comparator

The kind of comparison to make against a Qualification's value. You can compare a Qualification's value to an IntegerValue to see if it is LessThan, LessThanOrEqualTo, GreaterThan, GreaterThanOrEqualTo, EqualTo, or NotEqualTo the IntegerValue. You can compare it to a LocaleValue to see if it is EqualTo, or NotEqualTo the LocaleValue. You can check to see if the value is In or NotIn a set of IntegerValue or LocaleValue values. Lastly, a Qualification requirement can also test if a Qualification Exists or DoesNotExist in the user's profile, regardless of its value.

QualificationType

data QualificationType

The QualificationType data structure represents a Qualification type, a description of a property of a Worker that must match the requirements of a HIT for the Worker to be able to accept the HIT. The type also describes how a Worker can obtain a Qualification of that type, such as through a Qualification test.

See: qualificationType smart constructor.

qualificationType :: QualificationType

Creates a value of QualificationType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

qtCreationTime - The date and time the Qualification type was created.

qtTestDurationInSeconds - The amount of time, in seconds, given to a Worker to complete the Qualification test, beginning from the time the Worker requests the Qualification.

qtQualificationTypeStatus - The status of the Qualification type. A Qualification type's status determines if users can apply to receive a Qualification of this type, and if HITs can be created with requirements based on this type. Valid values are Active | Inactive.

qtAnswerKey - The answers to the Qualification test specified in the Test parameter.

qtTest - The questions for a Qualification test associated with this Qualification type that a user can take to obtain a Qualification of this type. This parameter must be specified if AnswerKey is present. A Qualification type cannot have both a specified Test parameter and an AutoGranted value of true.

qtQualificationTypeId - A unique identifier for the Qualification type. A Qualification type is given a Qualification type ID when you call the CreateQualificationType operation.

qtName - The name of the Qualification type. The type name is used to identify the type, and to find the type using a Qualification type search.

qtKeywords - One or more words or phrases that describe theQualification type, separated by commas. The Keywords make the type easier to find using a search.

qtAutoGranted - Specifies that requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Valid values are True | False.

qtAutoGrantedValue - The Qualification integer value to use for automatically granted Qualifications, if AutoGranted is true. This is 1 by default.

qtDescription - A long description for the Qualification type.

qtIsRequestable - Specifies whether the Qualification type is one that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test. This value is False for Qualifications assigned automatically by the system. Valid values are True | False.

qtRetryDelayInSeconds - The amount of time, in seconds, Workers must wait after taking the Qualification test before they can take it again. Workers can take a Qualification test multiple times if they were not granted the Qualification from a previous attempt, or if the test offers a gradient score and they want a better score. If not specified, retries are disabled and Workers can request a Qualification only once.

qtCreationTime :: Lens' QualificationType (Maybe UTCTime)

The date and time the Qualification type was created.

qtTestDurationInSeconds :: Lens' QualificationType (Maybe Integer)

The amount of time, in seconds, given to a Worker to complete the Qualification test, beginning from the time the Worker requests the Qualification.

qtQualificationTypeStatus :: Lens' QualificationType (Maybe QualificationTypeStatus)

The status of the Qualification type. A Qualification type's status determines if users can apply to receive a Qualification of this type, and if HITs can be created with requirements based on this type. Valid values are Active | Inactive.

qtAnswerKey :: Lens' QualificationType (Maybe Text)

The answers to the Qualification test specified in the Test parameter.

qtTest :: Lens' QualificationType (Maybe Text)

The questions for a Qualification test associated with this Qualification type that a user can take to obtain a Qualification of this type. This parameter must be specified if AnswerKey is present. A Qualification type cannot have both a specified Test parameter and an AutoGranted value of true.

qtQualificationTypeId :: Lens' QualificationType (Maybe Text)

A unique identifier for the Qualification type. A Qualification type is given a Qualification type ID when you call the CreateQualificationType operation.

qtName :: Lens' QualificationType (Maybe Text)

The name of the Qualification type. The type name is used to identify the type, and to find the type using a Qualification type search.

qtKeywords :: Lens' QualificationType (Maybe Text)

One or more words or phrases that describe theQualification type, separated by commas. The Keywords make the type easier to find using a search.

qtAutoGranted :: Lens' QualificationType (Maybe Bool)

Specifies that requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Valid values are True | False.

qtAutoGrantedValue :: Lens' QualificationType (Maybe Int)

The Qualification integer value to use for automatically granted Qualifications, if AutoGranted is true. This is 1 by default.

qtDescription :: Lens' QualificationType (Maybe Text)

A long description for the Qualification type.

qtIsRequestable :: Lens' QualificationType (Maybe Bool)

Specifies whether the Qualification type is one that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test. This value is False for Qualifications assigned automatically by the system. Valid values are True | False.

qtRetryDelayInSeconds :: Lens' QualificationType (Maybe Integer)

The amount of time, in seconds, Workers must wait after taking the Qualification test before they can take it again. Workers can take a Qualification test multiple times if they were not granted the Qualification from a previous attempt, or if the test offers a gradient score and they want a better score. If not specified, retries are disabled and Workers can request a Qualification only once.

ReviewActionDetail

data ReviewActionDetail

Both the AssignmentReviewReport and the HITReviewReport elements contains the ReviewActionDetail data structure. This structure is returned multiple times for each action specified in the Review Policy.

See: reviewActionDetail smart constructor.

reviewActionDetail :: ReviewActionDetail

Creates a value of ReviewActionDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

radStatus - The current disposition of the action: INTENDED, SUCCEEDED, FAILED, or CANCELLED.

radTargetId - The specific HITId or AssignmentID targeted by the action.

radActionId - The unique identifier for the action.

radTargetType - The type of object in TargetId.

radResult - A description of the outcome of the review.

radActionName - The nature of the action itself. The Review Policy is responsible for examining the HIT and Assignments, emitting results, and deciding which other actions will be necessary.

radCompleteTime - The date when the action was completed.

radErrorCode - Present only when the Results have a FAILED Status.

radStatus :: Lens' ReviewActionDetail (Maybe ReviewActionStatus)

The current disposition of the action: INTENDED, SUCCEEDED, FAILED, or CANCELLED.

radTargetId :: Lens' ReviewActionDetail (Maybe Text)

The specific HITId or AssignmentID targeted by the action.

radActionId :: Lens' ReviewActionDetail (Maybe Text)

The unique identifier for the action.

radTargetType :: Lens' ReviewActionDetail (Maybe Text)

The type of object in TargetId.

radResult :: Lens' ReviewActionDetail (Maybe Text)

A description of the outcome of the review.

radActionName :: Lens' ReviewActionDetail (Maybe Text)

The nature of the action itself. The Review Policy is responsible for examining the HIT and Assignments, emitting results, and deciding which other actions will be necessary.

radCompleteTime :: Lens' ReviewActionDetail (Maybe UTCTime)

The date when the action was completed.

radErrorCode :: Lens' ReviewActionDetail (Maybe Text)

Present only when the Results have a FAILED Status.

ReviewPolicy

data ReviewPolicy

HIT Review Policy data structures represent HIT review policies, which you specify when you create a HIT.

See: reviewPolicy smart constructor.

reviewPolicy :: ReviewPolicy

Creates a value of ReviewPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpPolicyName - Name of a Review Policy: SimplePlurality2011-09-01 or ScoreMyKnownAnswers2011-09-01

rpParameters - Name of the parameter from the Review policy.

rpPolicyName :: Lens' ReviewPolicy (Maybe Text)

Name of a Review Policy: SimplePlurality2011-09-01 or ScoreMyKnownAnswers2011-09-01

rpParameters :: Lens' ReviewPolicy [PolicyParameter]

Name of the parameter from the Review policy.

ReviewReport

data ReviewReport

Contains both ReviewResult and ReviewAction elements for a particular HIT.

See: reviewReport smart constructor.

reviewReport :: ReviewReport

Creates a value of ReviewReport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrReviewActions - A list of ReviewAction objects for each action specified in the Review Policy.

rrReviewResults - A list of ReviewResults objects for each action specified in the Review Policy.

rrReviewActions :: Lens' ReviewReport [ReviewActionDetail]

A list of ReviewAction objects for each action specified in the Review Policy.

rrReviewResults :: Lens' ReviewReport [ReviewResultDetail]

A list of ReviewResults objects for each action specified in the Review Policy.

ReviewResultDetail

data ReviewResultDetail

This data structure is returned multiple times for each result specified in the Review Policy.

See: reviewResultDetail smart constructor.

reviewResultDetail :: ReviewResultDetail

Creates a value of ReviewResultDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrdValue - The values of Key provided by the review policies you have selected.

rrdActionId - A unique identifier of the Review action result.

rrdSubjectType - The type of the object from the SubjectId field.

rrdKey - Key identifies the particular piece of reviewed information.

rrdQuestionId - Specifies the QuestionId the result is describing. Depending on whether the TargetType is a HIT or Assignment this results could specify multiple values. If TargetType is HIT and QuestionId is absent, then the result describes results of the HIT, including the HIT agreement score. If ObjectType is Assignment and QuestionId is absent, then the result describes the Worker's performance on the HIT.

rrdSubjectId - The HITID or AssignmentId about which this result was taken. Note that HIT-level Review Policies will often emit results about both the HIT itself and its Assignments, while Assignment-level review policies generally only emit results about the Assignment itself.

rrdValue :: Lens' ReviewResultDetail (Maybe Text)

The values of Key provided by the review policies you have selected.

rrdActionId :: Lens' ReviewResultDetail (Maybe Text)

A unique identifier of the Review action result.

rrdSubjectType :: Lens' ReviewResultDetail (Maybe Text)

The type of the object from the SubjectId field.

rrdKey :: Lens' ReviewResultDetail (Maybe Text)

Key identifies the particular piece of reviewed information.

rrdQuestionId :: Lens' ReviewResultDetail (Maybe Text)

Specifies the QuestionId the result is describing. Depending on whether the TargetType is a HIT or Assignment this results could specify multiple values. If TargetType is HIT and QuestionId is absent, then the result describes results of the HIT, including the HIT agreement score. If ObjectType is Assignment and QuestionId is absent, then the result describes the Worker's performance on the HIT.

rrdSubjectId :: Lens' ReviewResultDetail (Maybe Text)

The HITID or AssignmentId about which this result was taken. Note that HIT-level Review Policies will often emit results about both the HIT itself and its Assignments, while Assignment-level review policies generally only emit results about the Assignment itself.

WorkerBlock

data WorkerBlock

The WorkerBlock data structure represents a Worker who has been blocked. It has two elements: the WorkerId and the Reason for the block.

See: workerBlock smart constructor.

workerBlock :: WorkerBlock

Creates a value of WorkerBlock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

wbReason - A message explaining the reason the Worker was blocked.

wbWorkerId - The ID of the Worker who accepted the HIT.

wbReason :: Lens' WorkerBlock (Maybe Text)

A message explaining the reason the Worker was blocked.

wbWorkerId :: Lens' WorkerBlock (Maybe Text)

The ID of the Worker who accepted the HIT.