Network.AWS.MechanicalTurk.AcceptQualificationRequest

Description

The AcceptQualificationRequest operation approves a Worker's request for a Qualification.

Only the owner of the Qualification type can grant a Qualification request for that type.

A successful request for the AcceptQualificationRequest operation returns with no errors and an empty body.

Creating a Request

acceptQualificationRequest

Arguments

:: Text
aqrQualificationRequestId

-> AcceptQualificationRequest

Creates a value of AcceptQualificationRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aqrIntegerValue - The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.

aqrQualificationRequestId - The ID of the Qualification request, as returned by the GetQualificationRequests operation.

data AcceptQualificationRequest

See: acceptQualificationRequest smart constructor.

Request Lenses

aqrIntegerValue :: Lens' AcceptQualificationRequest (Maybe Int)

The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.

aqrQualificationRequestId :: Lens' AcceptQualificationRequest Text

The ID of the Qualification request, as returned by the GetQualificationRequests operation.

Destructuring the Response

acceptQualificationRequestResponse

Arguments

:: Int
aqrrsResponseStatus

-> AcceptQualificationRequestResponse

Creates a value of AcceptQualificationRequestResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aqrrsResponseStatus - -- | The response status code.

data AcceptQualificationRequestResponse

See: acceptQualificationRequestResponse smart constructor.

Response Lenses

aqrrsResponseStatus :: Lens' AcceptQualificationRequestResponse Int

- | The response status code.