Network.AWS.MechanicalTurk.ApproveAssignment

Description

The ApproveAssignment operation approves the results of a completed assignment.

Approving an assignment initiates two payments from the Requester's Amazon.com account

The Worker who submitted the results is paid the reward specified in the HIT.

Amazon Mechanical Turk fees are debited.

If the Requester's account does not have adequate funds for these payments, the call to ApproveAssignment returns an exception, and the approval is not processed. You can include an optional feedback message with the approval, which the Worker can see in the Status section of the web site.

You can also call this operation for assignments that were previous rejected and approve them by explicitly overriding the previous rejection. This only works on rejected assignments that were submitted within the previous 30 days and only if the assignment's related HIT has not been deleted.

Creating a Request

approveAssignment

Arguments

:: Text
aaAssignmentId

-> ApproveAssignment

Creates a value of ApproveAssignment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaOverrideRejection - A flag indicating that an assignment should be approved even if it was previously rejected. Defaults to False .

aaRequesterFeedback - A message for the Worker, which the Worker can see in the Status section of the web site.

aaAssignmentId - The ID of the assignment. The assignment must correspond to a HIT created by the Requester.

data ApproveAssignment

See: approveAssignment smart constructor.

Request Lenses

aaOverrideRejection :: Lens' ApproveAssignment (Maybe Bool)

A flag indicating that an assignment should be approved even if it was previously rejected. Defaults to False .

aaRequesterFeedback :: Lens' ApproveAssignment (Maybe Text)

A message for the Worker, which the Worker can see in the Status section of the web site.

aaAssignmentId :: Lens' ApproveAssignment Text

The ID of the assignment. The assignment must correspond to a HIT created by the Requester.

Destructuring the Response

approveAssignmentResponse

Arguments

:: Int
aarsResponseStatus

-> ApproveAssignmentResponse

Creates a value of ApproveAssignmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aarsResponseStatus - -- | The response status code.

data ApproveAssignmentResponse

See: approveAssignmentResponse smart constructor.

Response Lenses

aarsResponseStatus :: Lens' ApproveAssignmentResponse Int

- | The response status code.