Network.AWS.MechanicalTurk.AssociateQualificationWithWorker

Description

The AssociateQualificationWithWorker operation gives a Worker a Qualification. AssociateQualificationWithWorker does not require that the Worker submit a Qualification request. It gives the Qualification directly to the Worker.

You can only assign a Qualification of a Qualification type that you created (using the CreateQualificationType operation).

Creating a Request

associateQualificationWithWorker

Arguments

:: Text
aqwwQualificationTypeId

-> Text
aqwwWorkerId

-> AssociateQualificationWithWorker

Creates a value of AssociateQualificationWithWorker with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aqwwIntegerValue - The value of the Qualification to assign.

aqwwSendNotification - Specifies whether to send a notification email message to the Worker saying that the qualification was assigned to the Worker. Note: this is true by default.

aqwwQualificationTypeId - The ID of the Qualification type to use for the assigned Qualification.

aqwwWorkerId - The ID of the Worker to whom the Qualification is being assigned. Worker IDs are included with submitted HIT assignments and Qualification requests.

data AssociateQualificationWithWorker

See: associateQualificationWithWorker smart constructor.

Request Lenses

aqwwIntegerValue :: Lens' AssociateQualificationWithWorker (Maybe Int)

The value of the Qualification to assign.

aqwwSendNotification :: Lens' AssociateQualificationWithWorker (Maybe Bool)

Specifies whether to send a notification email message to the Worker saying that the qualification was assigned to the Worker. Note: this is true by default.

aqwwQualificationTypeId :: Lens' AssociateQualificationWithWorker Text

The ID of the Qualification type to use for the assigned Qualification.

aqwwWorkerId :: Lens' AssociateQualificationWithWorker Text

The ID of the Worker to whom the Qualification is being assigned. Worker IDs are included with submitted HIT assignments and Qualification requests.

Destructuring the Response

associateQualificationWithWorkerResponse

Arguments

:: Int
aqwwrsResponseStatus

-> AssociateQualificationWithWorkerResponse

Creates a value of AssociateQualificationWithWorkerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aqwwrsResponseStatus - -- | The response status code.

data AssociateQualificationWithWorkerResponse

See: associateQualificationWithWorkerResponse smart constructor.

Response Lenses

aqwwrsResponseStatus :: Lens' AssociateQualificationWithWorkerResponse Int

- | The response status code.