Network.AWS.MechanicalTurk.CreateHITWithHITType

Description

The CreateHITWithHITType operation creates a new Human Intelligence Task (HIT) using an existing HITTypeID generated by the CreateHITType operation.

This is an alternative way to create HITs from the CreateHIT operation. This is the recommended best practice for Requesters who are creating large numbers of HITs.

CreateHITWithHITType also supports several ways to provide question data: by providing a value for the Question parameter that fully specifies the contents of the HIT, or by providing a HitLayoutId and associated HitLayoutParameters .

Creating a Request

createHITWithHITType

Arguments

:: Text
chitwhittHITTypeId

-> Integer
chitwhittLifetimeInSeconds

-> CreateHITWithHITType

Creates a value of CreateHITWithHITType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chitwhittHITReviewPolicy - The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

chitwhittUniqueRequestToken - A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

chitwhittRequesterAnnotation - An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT. The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester. The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

chitwhittMaxAssignments - The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

chitwhittHITLayoutId - The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters. Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

chitwhittHITLayoutParameters - If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

chitwhittQuestion - The data the person completing the HIT uses to produce the results. Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace. Either a Question parameter or a HITLayoutId parameter must be provided.

chitwhittAssignmentReviewPolicy - The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

chitwhittHITTypeId - The HIT type ID you want to create this HIT with.

chitwhittLifetimeInSeconds - An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

data CreateHITWithHITType

See: createHITWithHITType smart constructor.

Request Lenses

chitwhittHITReviewPolicy :: Lens' CreateHITWithHITType (Maybe ReviewPolicy)

The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

chitwhittUniqueRequestToken :: Lens' CreateHITWithHITType (Maybe Text)

A unique identifier for this request which allows you to retry the call on error without creating duplicate HITs. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return a AWS.MechanicalTurk.HitAlreadyExists error with a message containing the HITId.

chitwhittRequesterAnnotation :: Lens' CreateHITWithHITType (Maybe Text)

An arbitrary data field. The RequesterAnnotation parameter lets your application attach arbitrary data to the HIT for tracking purposes. For example, this parameter could be an identifier internal to the Requester's application that corresponds with the HIT. The RequesterAnnotation parameter for a HIT is only visible to the Requester who created the HIT. It is not shown to the Worker, or any other Requester. The RequesterAnnotation parameter may be different for each HIT you submit. It does not affect how your HITs are grouped.

chitwhittMaxAssignments :: Lens' CreateHITWithHITType (Maybe Int)

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

chitwhittHITLayoutId :: Lens' CreateHITWithHITType (Maybe Text)

The HITLayoutId allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters. Constraints: Either a Question parameter or a HITLayoutId parameter must be provided.

chitwhittHITLayoutParameters :: Lens' CreateHITWithHITType [HITLayoutParameter]

If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure. For more information, see HITLayout.

chitwhittQuestion :: Lens' CreateHITWithHITType (Maybe Text)

The data the person completing the HIT uses to produce the results. Constraints: Must be a QuestionForm data structure, an ExternalQuestion data structure, or an HTMLQuestion data structure. The XML question data must not be larger than 64 kilobytes (65,535 bytes) in size, including whitespace. Either a Question parameter or a HITLayoutId parameter must be provided.

chitwhittAssignmentReviewPolicy :: Lens' CreateHITWithHITType (Maybe ReviewPolicy)

The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.

chitwhittHITTypeId :: Lens' CreateHITWithHITType Text

The HIT type ID you want to create this HIT with.

chitwhittLifetimeInSeconds :: Lens' CreateHITWithHITType Integer

An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.

Destructuring the Response

createHITWithHITTypeResponse

Arguments

:: Int
chitwhittrsResponseStatus

-> CreateHITWithHITTypeResponse

Creates a value of CreateHITWithHITTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chitwhittrsHIT - Contains the newly created HIT data. For a description of the HIT data structure as it appears in responses, see the HIT Data Structure documentation.

chitwhittrsResponseStatus - -- | The response status code.

data CreateHITWithHITTypeResponse

See: createHITWithHITTypeResponse smart constructor.

Response Lenses

chitwhittrsHIT :: Lens' CreateHITWithHITTypeResponse (Maybe HIT)

Contains the newly created HIT data. For a description of the HIT data structure as it appears in responses, see the HIT Data Structure documentation.

chitwhittrsResponseStatus :: Lens' CreateHITWithHITTypeResponse Int

- | The response status code.