Network.AWS.MechanicalTurk.DisassociateQualificationFromWorker

Description

The DisassociateQualificationFromWorker revokes a previously granted Qualification from a user.

You can provide a text message explaining why the Qualification was revoked. The user who had the Qualification can see this message.

Creating a Request

disassociateQualificationFromWorker

Arguments

:: Text
dqfwWorkerId

-> Text
dqfwQualificationTypeId

-> DisassociateQualificationFromWorker

Creates a value of DisassociateQualificationFromWorker with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dqfwReason - A text message that explains why the Qualification was revoked. The user who had the Qualification sees this message.

dqfwWorkerId - The ID of the Worker who possesses the Qualification to be revoked.

dqfwQualificationTypeId - The ID of the Qualification type of the Qualification to be revoked.

data DisassociateQualificationFromWorker

See: disassociateQualificationFromWorker smart constructor.

Request Lenses

dqfwReason :: Lens' DisassociateQualificationFromWorker (Maybe Text)

A text message that explains why the Qualification was revoked. The user who had the Qualification sees this message.

dqfwWorkerId :: Lens' DisassociateQualificationFromWorker Text

The ID of the Worker who possesses the Qualification to be revoked.

dqfwQualificationTypeId :: Lens' DisassociateQualificationFromWorker Text

The ID of the Qualification type of the Qualification to be revoked.

Destructuring the Response

disassociateQualificationFromWorkerResponse

Arguments

:: Int
dqfwrsResponseStatus

-> DisassociateQualificationFromWorkerResponse

Creates a value of DisassociateQualificationFromWorkerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dqfwrsResponseStatus - -- | The response status code.

data DisassociateQualificationFromWorkerResponse

See: disassociateQualificationFromWorkerResponse smart constructor.

Response Lenses

dqfwrsResponseStatus :: Lens' DisassociateQualificationFromWorkerResponse Int

- | The response status code.