Network.AWS.MechanicalTurk.GetAssignment

Description

The GetAssignment operation retrieves the details of the specified Assignment.

Creating a Request

getAssignment

Arguments

:: Text
gaAssignmentId

-> GetAssignment

Creates a value of GetAssignment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaAssignmentId - The ID of the Assignment to be retrieved.

data GetAssignment

See: getAssignment smart constructor.

Request Lenses

gaAssignmentId :: Lens' GetAssignment Text

The ID of the Assignment to be retrieved.

Destructuring the Response

getAssignmentResponse

Arguments

:: Int
garsResponseStatus

-> GetAssignmentResponse

Creates a value of GetAssignmentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garsHIT - The HIT associated with this assignment. The response includes one HIT element.

garsAssignment - The assignment. The response includes one Assignment element.

garsResponseStatus - -- | The response status code.

data GetAssignmentResponse

See: getAssignmentResponse smart constructor.

Response Lenses

garsHIT :: Lens' GetAssignmentResponse (Maybe HIT)

The HIT associated with this assignment. The response includes one HIT element.

garsAssignment :: Lens' GetAssignmentResponse (Maybe Assignment)

The assignment. The response includes one Assignment element.

garsResponseStatus :: Lens' GetAssignmentResponse Int

- | The response status code.