Network.AWS.MechanicalTurk.GetFileUploadURL

Description

The GetFileUploadURL operation generates and returns a temporary URL. You use the temporary URL to retrieve a file uploaded by a Worker as an answer to a FileUploadAnswer question for a HIT. The temporary URL is generated the instant the GetFileUploadURL operation is called, and is valid for 60 seconds. You can get a temporary file upload URL any time until the HIT is disposed. After the HIT is disposed, any uploaded files are deleted, and cannot be retrieved.

Creating a Request

getFileUploadURL

Arguments

:: Text
gfuuAssignmentId

-> Text
gfuuQuestionIdentifier

-> GetFileUploadURL

Creates a value of GetFileUploadURL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfuuAssignmentId - The ID of the assignment that contains the question with a FileUploadAnswer.

gfuuQuestionIdentifier - The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.

data GetFileUploadURL

See: getFileUploadURL smart constructor.

Request Lenses

gfuuAssignmentId :: Lens' GetFileUploadURL Text

The ID of the assignment that contains the question with a FileUploadAnswer.

gfuuQuestionIdentifier :: Lens' GetFileUploadURL Text

The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.

Destructuring the Response

getFileUploadURLResponse

Arguments

:: Int
gfuursResponseStatus

-> GetFileUploadURLResponse

Creates a value of GetFileUploadURLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gfuursFileUploadURL - A temporary URL for the file that the Worker uploaded for the answer.

gfuursResponseStatus - -- | The response status code.

data GetFileUploadURLResponse

See: getFileUploadURLResponse smart constructor.

Response Lenses

gfuursFileUploadURL :: Lens' GetFileUploadURLResponse (Maybe Text)

A temporary URL for the file that the Worker uploaded for the answer.

gfuursResponseStatus :: Lens' GetFileUploadURLResponse Int

- | The response status code.