Network.AWS.MechanicalTurk.GetQualificationScore

Description

The GetQualificationScore operation returns the value of a Worker's Qualification for a given Qualification type.

To get a Worker's Qualification, you must know the Worker's ID. The Worker's ID is included in the assignment data returned by the ListAssignmentsForHIT operation.

Only the owner of a Qualification type can query the value of a Worker's Qualification of that type.

Creating a Request

getQualificationScore

Arguments

:: Text
gqsQualificationTypeId

-> Text
gqsWorkerId

-> GetQualificationScore

Creates a value of GetQualificationScore with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gqsQualificationTypeId - The ID of the QualificationType.

gqsWorkerId - The ID of the Worker whose Qualification is being updated.

data GetQualificationScore

See: getQualificationScore smart constructor.

Request Lenses

gqsQualificationTypeId :: Lens' GetQualificationScore Text

The ID of the QualificationType.

gqsWorkerId :: Lens' GetQualificationScore Text

The ID of the Worker whose Qualification is being updated.

Destructuring the Response

getQualificationScoreResponse

Arguments

:: Int
gqsrsResponseStatus

-> GetQualificationScoreResponse

Creates a value of GetQualificationScoreResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gqsrsQualification - The Qualification data structure of the Qualification assigned to a user, including the Qualification type and the value (score).

gqsrsResponseStatus - -- | The response status code.

data GetQualificationScoreResponse

See: getQualificationScoreResponse smart constructor.

Response Lenses

gqsrsQualification :: Lens' GetQualificationScoreResponse (Maybe Qualification)

The Qualification data structure of the Qualification assigned to a user, including the Qualification type and the value (score).

gqsrsResponseStatus :: Lens' GetQualificationScoreResponse Int

- | The response status code.