Network.AWS.MechanicalTurk.ListAssignmentsForHIT

Description

The ListAssignmentsForHIT operation retrieves completed assignments for a HIT. You can use this operation to retrieve the results for a HIT.

You can get assignments for a HIT at any time, even if the HIT is not yet Reviewable. If a HIT requested multiple assignments, and has received some results but has not yet become Reviewable, you can still retrieve the partial results with this operation.

Use the AssignmentStatus parameter to control which set of assignments for a HIT are returned. The ListAssignmentsForHIT operation can return submitted assignments awaiting approval, or it can return assignments that have already been approved or rejected. You can set AssignmentStatus=Approved,Rejected to get assignments that have already been approved and rejected together in one result set.

Only the Requester who created the HIT can retrieve the assignments for that HIT.

Results are sorted and divided into numbered pages and the operation returns a single page of results. You can use the parameters of the operation to control sorting and pagination.

Creating a Request

listAssignmentsForHIT

Arguments

:: Text
lafhitHITId

-> ListAssignmentsForHIT

Creates a value of ListAssignmentsForHIT with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lafhitAssignmentStatuses - The status of the assignments to return: Submitted | Approved | Rejected

lafhitNextToken - Pagination token

lafhitMaxResults - Undocumented member.

lafhitHITId - The ID of the HIT.

data ListAssignmentsForHIT

See: listAssignmentsForHIT smart constructor.

Request Lenses

lafhitAssignmentStatuses :: Lens' ListAssignmentsForHIT [AssignmentStatus]

The status of the assignments to return: Submitted | Approved | Rejected

lafhitNextToken :: Lens' ListAssignmentsForHIT (Maybe Text)

Pagination token

lafhitMaxResults :: Lens' ListAssignmentsForHIT (Maybe Natural)

Undocumented member.

lafhitHITId :: Lens' ListAssignmentsForHIT Text

The ID of the HIT.

Destructuring the Response

listAssignmentsForHITResponse

Arguments

:: Int
lafhitrsResponseStatus

-> ListAssignmentsForHITResponse

Creates a value of ListAssignmentsForHITResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lafhitrsNextToken - Undocumented member.

lafhitrsNumResults - The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

lafhitrsAssignments - The collection of Assignment data structures returned by this call.

lafhitrsResponseStatus - -- | The response status code.

data ListAssignmentsForHITResponse

See: listAssignmentsForHITResponse smart constructor.

Response Lenses

lafhitrsNextToken :: Lens' ListAssignmentsForHITResponse (Maybe Text)

Undocumented member.

lafhitrsNumResults :: Lens' ListAssignmentsForHITResponse (Maybe Int)

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

lafhitrsAssignments :: Lens' ListAssignmentsForHITResponse [Assignment]

The collection of Assignment data structures returned by this call.

lafhitrsResponseStatus :: Lens' ListAssignmentsForHITResponse Int

- | The response status code.