Network.AWS.MechanicalTurk.ListBonusPayments

Description

The ListBonusPayments operation retrieves the amounts of bonuses you have paid to Workers for a given HIT or assignment.

Creating a Request

listBonusPayments :: ListBonusPayments

Creates a value of ListBonusPayments with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbpNextToken - Pagination token

lbpHITId - The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified

lbpAssignmentId - The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified

lbpMaxResults - Undocumented member.

data ListBonusPayments

See: listBonusPayments smart constructor.

Request Lenses

lbpNextToken :: Lens' ListBonusPayments (Maybe Text)

Pagination token

lbpHITId :: Lens' ListBonusPayments (Maybe Text)

The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified

lbpAssignmentId :: Lens' ListBonusPayments (Maybe Text)

The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified

lbpMaxResults :: Lens' ListBonusPayments (Maybe Natural)

Undocumented member.

Destructuring the Response

listBonusPaymentsResponse

Arguments

:: Int
lbprsResponseStatus

-> ListBonusPaymentsResponse

Creates a value of ListBonusPaymentsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbprsBonusPayments - A successful request to the ListBonusPayments operation returns a list of BonusPayment objects.

lbprsNextToken - Undocumented member.

lbprsNumResults - The number of bonus payments on this page in the filtered results list, equivalent to the number of bonus payments being returned by this call.

lbprsResponseStatus - -- | The response status code.

data ListBonusPaymentsResponse

See: listBonusPaymentsResponse smart constructor.

Response Lenses

lbprsBonusPayments :: Lens' ListBonusPaymentsResponse [BonusPayment]

A successful request to the ListBonusPayments operation returns a list of BonusPayment objects.

lbprsNextToken :: Lens' ListBonusPaymentsResponse (Maybe Text)

Undocumented member.

lbprsNumResults :: Lens' ListBonusPaymentsResponse (Maybe Int)

The number of bonus payments on this page in the filtered results list, equivalent to the number of bonus payments being returned by this call.

lbprsResponseStatus :: Lens' ListBonusPaymentsResponse Int

- | The response status code.