Network.AWS.MechanicalTurk.ListHITsForQualificationType

Description

The ListHITsForQualificationType operation returns the HITs that use the given Qualification type for a Qualification requirement. The operation returns HITs of any status, except for HITs that have been deleted with the DeleteHIT operation or that have been auto-deleted.

Creating a Request

listHITsForQualificationType

Arguments

:: Text
lhitfqtQualificationTypeId

-> ListHITsForQualificationType

Creates a value of ListHITsForQualificationType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhitfqtNextToken - Pagination Token

lhitfqtMaxResults - Limit the number of results returned.

lhitfqtQualificationTypeId - The ID of the Qualification type to use when querying HITs.

data ListHITsForQualificationType

See: listHITsForQualificationType smart constructor.

Request Lenses

lhitfqtNextToken :: Lens' ListHITsForQualificationType (Maybe Text)

Pagination Token

lhitfqtMaxResults :: Lens' ListHITsForQualificationType (Maybe Natural)

Limit the number of results returned.

lhitfqtQualificationTypeId :: Lens' ListHITsForQualificationType Text

The ID of the Qualification type to use when querying HITs.

Destructuring the Response

listHITsForQualificationTypeResponse

Arguments

:: Int
lhitfqtrsResponseStatus

-> ListHITsForQualificationTypeResponse

Creates a value of ListHITsForQualificationTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhitfqtrsNextToken - Undocumented member.

lhitfqtrsNumResults - The number of HITs on this page in the filtered results list, equivalent to the number of HITs being returned by this call.

lhitfqtrsHITs - The list of HIT elements returned by the query.

lhitfqtrsResponseStatus - -- | The response status code.

data ListHITsForQualificationTypeResponse

See: listHITsForQualificationTypeResponse smart constructor.

Response Lenses

lhitfqtrsNextToken :: Lens' ListHITsForQualificationTypeResponse (Maybe Text)

Undocumented member.

lhitfqtrsNumResults :: Lens' ListHITsForQualificationTypeResponse (Maybe Int)

The number of HITs on this page in the filtered results list, equivalent to the number of HITs being returned by this call.

lhitfqtrsHITs :: Lens' ListHITsForQualificationTypeResponse [HIT]

The list of HIT elements returned by the query.

lhitfqtrsResponseStatus :: Lens' ListHITsForQualificationTypeResponse Int

- | The response status code.