Network.AWS.MechanicalTurk.ListQualificationTypes

Description

The ListQualificationRequests operation retrieves requests for Qualifications of a particular Qualification type. The owner of the Qualification type calls this operation to poll for pending requests, and accepts them using the AcceptQualification operation.

Creating a Request

listQualificationTypes

Arguments

:: Bool
lqtMustBeRequestable

-> ListQualificationTypes

Creates a value of ListQualificationTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqtMustBeOwnedByCaller - Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.

lqtNextToken - Undocumented member.

lqtQuery - A text query against all of the searchable attributes of Qualification types.

lqtMaxResults - The maximum number of results to return in a single call.

lqtMustBeRequestable - Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered. Valid values are True | False.

data ListQualificationTypes

See: listQualificationTypes smart constructor.

Request Lenses

lqtMustBeOwnedByCaller :: Lens' ListQualificationTypes (Maybe Bool)

Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.

lqtNextToken :: Lens' ListQualificationTypes (Maybe Text)

Undocumented member.

lqtQuery :: Lens' ListQualificationTypes (Maybe Text)

A text query against all of the searchable attributes of Qualification types.

lqtMaxResults :: Lens' ListQualificationTypes (Maybe Natural)

The maximum number of results to return in a single call.

lqtMustBeRequestable :: Lens' ListQualificationTypes Bool

Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered. Valid values are True | False.

Destructuring the Response

listQualificationTypesResponse

Arguments

:: Int
lqtrsResponseStatus

-> ListQualificationTypesResponse

Creates a value of ListQualificationTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lqtrsQualificationTypes - The list of QualificationType elements returned by the query.

lqtrsNextToken - Undocumented member.

lqtrsNumResults - The number of Qualification types on this page in the filtered results list, equivalent to the number of types this operation returns.

lqtrsResponseStatus - -- | The response status code.

data ListQualificationTypesResponse

See: listQualificationTypesResponse smart constructor.

Response Lenses

lqtrsQualificationTypes :: Lens' ListQualificationTypesResponse [QualificationType]

The list of QualificationType elements returned by the query.

lqtrsNextToken :: Lens' ListQualificationTypesResponse (Maybe Text)

Undocumented member.

lqtrsNumResults :: Lens' ListQualificationTypesResponse (Maybe Int)

The number of Qualification types on this page in the filtered results list, equivalent to the number of types this operation returns.

lqtrsResponseStatus :: Lens' ListQualificationTypesResponse Int

- | The response status code.