Network.AWS.MechanicalTurk.ListReviewPolicyResultsForHIT

Description

The ListReviewPolicyResultsForHIT operation retrieves the computed results and the actions taken in the course of executing your Review Policies for a given HIT. For information about how to specify Review Policies when you call CreateHIT, see Review Policies. The ListReviewPolicyResultsForHIT operation can return results for both Assignment-level and HIT-level review results.

Creating a Request

listReviewPolicyResultsForHIT

Arguments

:: Text
lrprfhitHITId

-> ListReviewPolicyResultsForHIT

Creates a value of ListReviewPolicyResultsForHIT with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrprfhitRetrieveResults - Specify if the operation should retrieve a list of the results computed by the Review Policies.

lrprfhitPolicyLevels - The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels. For a list of all the described policies, see Review Policies.

lrprfhitRetrieveActions - Specify if the operation should retrieve a list of the actions taken executing the Review Policies and their outcomes.

lrprfhitNextToken - Pagination token

lrprfhitMaxResults - Limit the number of results returned.

lrprfhitHITId - The unique identifier of the HIT to retrieve review results for.

data ListReviewPolicyResultsForHIT

See: listReviewPolicyResultsForHIT smart constructor.

Request Lenses

lrprfhitRetrieveResults :: Lens' ListReviewPolicyResultsForHIT (Maybe Bool)

Specify if the operation should retrieve a list of the results computed by the Review Policies.

lrprfhitPolicyLevels :: Lens' ListReviewPolicyResultsForHIT [ReviewPolicyLevel]

The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels. For a list of all the described policies, see Review Policies.

lrprfhitRetrieveActions :: Lens' ListReviewPolicyResultsForHIT (Maybe Bool)

Specify if the operation should retrieve a list of the actions taken executing the Review Policies and their outcomes.

lrprfhitNextToken :: Lens' ListReviewPolicyResultsForHIT (Maybe Text)

Pagination token

lrprfhitMaxResults :: Lens' ListReviewPolicyResultsForHIT (Maybe Natural)

Limit the number of results returned.

lrprfhitHITId :: Lens' ListReviewPolicyResultsForHIT Text

The unique identifier of the HIT to retrieve review results for.

Destructuring the Response

listReviewPolicyResultsForHITResponse

Arguments

:: Int
lrprfhitrsResponseStatus

-> ListReviewPolicyResultsForHITResponse

Creates a value of ListReviewPolicyResultsForHITResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrprfhitrsHITReviewPolicy - The name of the HIT-level Review Policy. This contains only the PolicyName element.

lrprfhitrsHITReviewReport - Contains both ReviewResult and ReviewAction elements for a particular HIT.

lrprfhitrsNextToken - Undocumented member.

lrprfhitrsAssignmentReviewReport - Contains both ReviewResult and ReviewAction elements for an Assignment.

lrprfhitrsHITId - The HITId of the HIT for which results have been returned.

lrprfhitrsAssignmentReviewPolicy - The name of the Assignment-level Review Policy. This contains only the PolicyName element.

lrprfhitrsResponseStatus - -- | The response status code.

data ListReviewPolicyResultsForHITResponse

See: listReviewPolicyResultsForHITResponse smart constructor.

Response Lenses

lrprfhitrsHITReviewPolicy :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewPolicy)

The name of the HIT-level Review Policy. This contains only the PolicyName element.

lrprfhitrsHITReviewReport :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewReport)

Contains both ReviewResult and ReviewAction elements for a particular HIT.

lrprfhitrsNextToken :: Lens' ListReviewPolicyResultsForHITResponse (Maybe Text)

Undocumented member.

lrprfhitrsAssignmentReviewReport :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewReport)

Contains both ReviewResult and ReviewAction elements for an Assignment.

lrprfhitrsHITId :: Lens' ListReviewPolicyResultsForHITResponse (Maybe Text)

The HITId of the HIT for which results have been returned.

lrprfhitrsAssignmentReviewPolicy :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewPolicy)

The name of the Assignment-level Review Policy. This contains only the PolicyName element.

lrprfhitrsResponseStatus :: Lens' ListReviewPolicyResultsForHITResponse Int

- | The response status code.