Network.AWS.MechanicalTurk.ListReviewableHITs

Description

The ListReviewableHITs operation retrieves the HITs with Status equal to Reviewable or Status equal to Reviewing that belong to the Requester calling the operation.

Creating a Request

listReviewableHITs :: ListReviewableHITs

Creates a value of ListReviewableHITs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrhitStatus - Can be either Reviewable or Reviewing . Reviewable is the default value.

lrhitHITTypeId - The ID of the HIT type of the HITs to consider for the query. If not specified, all HITs for the Reviewer are considered

lrhitNextToken - Pagination Token

lrhitMaxResults - Limit the number of results returned.

data ListReviewableHITs

See: listReviewableHITs smart constructor.

Request Lenses

lrhitStatus :: Lens' ListReviewableHITs (Maybe ReviewableHITStatus)

Can be either Reviewable or Reviewing . Reviewable is the default value.

lrhitHITTypeId :: Lens' ListReviewableHITs (Maybe Text)

The ID of the HIT type of the HITs to consider for the query. If not specified, all HITs for the Reviewer are considered

lrhitNextToken :: Lens' ListReviewableHITs (Maybe Text)

Pagination Token

lrhitMaxResults :: Lens' ListReviewableHITs (Maybe Natural)

Limit the number of results returned.

Destructuring the Response

listReviewableHITsResponse

Arguments

:: Int
lrhitrsResponseStatus

-> ListReviewableHITsResponse

Creates a value of ListReviewableHITsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrhitrsNextToken - Undocumented member.

lrhitrsNumResults - The number of HITs on this page in the filtered results list, equivalent to the number of HITs being returned by this call.

lrhitrsHITs - The list of HIT elements returned by the query.

lrhitrsResponseStatus - -- | The response status code.

data ListReviewableHITsResponse

See: listReviewableHITsResponse smart constructor.

Response Lenses

lrhitrsNextToken :: Lens' ListReviewableHITsResponse (Maybe Text)

Undocumented member.

lrhitrsNumResults :: Lens' ListReviewableHITsResponse (Maybe Int)

The number of HITs on this page in the filtered results list, equivalent to the number of HITs being returned by this call.

lrhitrsHITs :: Lens' ListReviewableHITsResponse [HIT]

The list of HIT elements returned by the query.

lrhitrsResponseStatus :: Lens' ListReviewableHITsResponse Int

- | The response status code.