Network.AWS.MechanicalTurk.ListWorkerBlocks

Description

The ListWorkersBlocks operation retrieves a list of Workers who are blocked from working on your HITs.

Creating a Request

listWorkerBlocks :: ListWorkerBlocks

Creates a value of ListWorkerBlocks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lwbNextToken - Pagination token

lwbMaxResults - Undocumented member.

data ListWorkerBlocks

See: listWorkerBlocks smart constructor.

Request Lenses

lwbNextToken :: Lens' ListWorkerBlocks (Maybe Text)

Pagination token

lwbMaxResults :: Lens' ListWorkerBlocks (Maybe Natural)

Undocumented member.

Destructuring the Response

listWorkerBlocksResponse

Arguments

:: Int
lwbrsResponseStatus

-> ListWorkerBlocksResponse

Creates a value of ListWorkerBlocksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lwbrsWorkerBlocks - The list of WorkerBlocks, containing the collection of Worker IDs and reasons for blocking.

lwbrsNextToken - Undocumented member.

lwbrsNumResults - The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

lwbrsResponseStatus - -- | The response status code.

data ListWorkerBlocksResponse

See: listWorkerBlocksResponse smart constructor.

Response Lenses

lwbrsWorkerBlocks :: Lens' ListWorkerBlocksResponse [WorkerBlock]

The list of WorkerBlocks, containing the collection of Worker IDs and reasons for blocking.

lwbrsNextToken :: Lens' ListWorkerBlocksResponse (Maybe Text)

Undocumented member.

lwbrsNumResults :: Lens' ListWorkerBlocksResponse (Maybe Int)

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

lwbrsResponseStatus :: Lens' ListWorkerBlocksResponse Int

- | The response status code.