Network.AWS.MechanicalTurk.NotifyWorkers

Description

The NotifyWorkers operation sends an email to one or more Workers that you specify with the Worker ID. You can specify up to 100 Worker IDs to send the same message with a single call to the NotifyWorkers operation. The NotifyWorkers operation will send a notification email to a Worker only if you have previously approved or rejected work from the Worker.

Creating a Request

notifyWorkers

Arguments

:: Text
nwSubject

-> Text
nwMessageText

-> NotifyWorkers

Creates a value of NotifyWorkers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nwSubject - The subject line of the email message to send. Can include up to 200 characters.

nwMessageText - The text of the email message to send. Can include up to 4,096 characters

nwWorkerIds - A list of Worker IDs you wish to notify. You can notify upto 100 Workers at a time.

data NotifyWorkers

See: notifyWorkers smart constructor.

Request Lenses

nwSubject :: Lens' NotifyWorkers Text

The subject line of the email message to send. Can include up to 200 characters.

nwMessageText :: Lens' NotifyWorkers Text

The text of the email message to send. Can include up to 4,096 characters

nwWorkerIds :: Lens' NotifyWorkers [Text]

A list of Worker IDs you wish to notify. You can notify upto 100 Workers at a time.

Destructuring the Response

notifyWorkersResponse

Arguments

:: Int
nwrsResponseStatus

-> NotifyWorkersResponse

Creates a value of NotifyWorkersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nwrsNotifyWorkersFailureStatuses - When MTurk sends notifications to the list of Workers, it returns back any failures it encounters in this list of NotifyWorkersFailureStatus objects.

nwrsResponseStatus - -- | The response status code.

data NotifyWorkersResponse

See: notifyWorkersResponse smart constructor.

Response Lenses

nwrsNotifyWorkersFailureStatuses :: Lens' NotifyWorkersResponse [NotifyWorkersFailureStatus]

When MTurk sends notifications to the list of Workers, it returns back any failures it encounters in this list of NotifyWorkersFailureStatus objects.

nwrsResponseStatus :: Lens' NotifyWorkersResponse Int

- | The response status code.