Network.AWS.MechanicalTurk.SendBonus

Description

The SendBonus operation issues a payment of money from your account to a Worker. This payment happens separately from the reward you pay to the Worker when you approve the Worker's assignment. The SendBonus operation requires the Worker's ID and the assignment ID as parameters to initiate payment of the bonus. You must include a message that explains the reason for the bonus payment, as the Worker may not be expecting the payment. Amazon Mechanical Turk collects a fee for bonus payments, similar to the HIT listing fee. This operation fails if your account does not have enough funds to pay for both the bonus and the fees.

Creating a Request

sendBonus

Arguments

:: Text
sbWorkerId

-> Text
sbBonusAmount

-> Text
sbAssignmentId

-> SendBonus

Creates a value of SendBonus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sbUniqueRequestToken - A unique identifier for this request, which allows you to retry the call on error without granting multiple bonuses. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the bonus already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return an error with a message containing the request ID.

sbReason - A message that explains the reason for the bonus payment. The Worker receiving the bonus can see this message.

sbWorkerId - The ID of the Worker being paid the bonus.

sbBonusAmount - The Bonus amount is a US Dollar amount specified using a string (for example, "5" represents $5.00 USD and "101.42" represents $101.42 USD). Do not include currency symbols or currency codes.

sbAssignmentId - The ID of the assignment for which this bonus is paid.

data SendBonus

See: sendBonus smart constructor.

Request Lenses

sbUniqueRequestToken :: Lens' SendBonus (Maybe Text)

A unique identifier for this request, which allows you to retry the call on error without granting multiple bonuses. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the bonus already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return an error with a message containing the request ID.

sbReason :: Lens' SendBonus (Maybe Text)

A message that explains the reason for the bonus payment. The Worker receiving the bonus can see this message.

sbWorkerId :: Lens' SendBonus Text

The ID of the Worker being paid the bonus.

sbBonusAmount :: Lens' SendBonus Text

The Bonus amount is a US Dollar amount specified using a string (for example, "5" represents $5.00 USD and "101.42" represents $101.42 USD). Do not include currency symbols or currency codes.

sbAssignmentId :: Lens' SendBonus Text

The ID of the assignment for which this bonus is paid.

Destructuring the Response

sendBonusResponse

Arguments

:: Int
sbrsResponseStatus

-> SendBonusResponse

Creates a value of SendBonusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sbrsResponseStatus - -- | The response status code.

data SendBonusResponse

See: sendBonusResponse smart constructor.

Response Lenses

sbrsResponseStatus :: Lens' SendBonusResponse Int

- | The response status code.