Network.AWS.MechanicalTurk.UpdateHITReviewStatus

Description

The UpdateHITReviewStatus operation updates the status of a HIT. If the status is Reviewable, this operation can update the status to Reviewing, or it can revert a Reviewing HIT back to the Reviewable status.

Creating a Request

updateHITReviewStatus

Arguments

:: Text
uhitrsHITId

-> UpdateHITReviewStatus

Creates a value of UpdateHITReviewStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uhitrsRevert - Specifies how to update the HIT status. Default is False . * Setting this to false will only transition a HIT from Reviewable to Reviewing * Setting this to true will only transition a HIT from Reviewing to Reviewable

uhitrsHITId - The ID of the HIT to update.

data UpdateHITReviewStatus

See: updateHITReviewStatus smart constructor.

Request Lenses

uhitrsRevert :: Lens' UpdateHITReviewStatus (Maybe Bool)

Specifies how to update the HIT status. Default is False . * Setting this to false will only transition a HIT from Reviewable to Reviewing * Setting this to true will only transition a HIT from Reviewing to Reviewable

uhitrsHITId :: Lens' UpdateHITReviewStatus Text

The ID of the HIT to update.

Destructuring the Response

updateHITReviewStatusResponse

Arguments

:: Int
uhitrsrsResponseStatus

-> UpdateHITReviewStatusResponse

Creates a value of UpdateHITReviewStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uhitrsrsResponseStatus - -- | The response status code.

data UpdateHITReviewStatusResponse

See: updateHITReviewStatusResponse smart constructor.

Response Lenses

uhitrsrsResponseStatus :: Lens' UpdateHITReviewStatusResponse Int

- | The response status code.