Network.AWS.MechanicalTurk.UpdateNotificationSettings

Description

The UpdateNotificationSettings operation creates, updates, disables or re-enables notifications for a HIT type. If you call the UpdateNotificationSettings operation for a HIT type that already has a notification specification, the operation replaces the old specification with a new one. You can call the UpdateNotificationSettings operation to enable or disable notifications for the HIT type, without having to modify the notification specification itself by providing updates to the Active status without specifying a new notification specification. To change the Active status of a HIT type's notifications, the HIT type must already have a notification specification, or one must be provided in the same call to UpdateNotificationSettings .

Creating a Request

updateNotificationSettings

Arguments

:: Text
unsHITTypeId

-> UpdateNotificationSettings

Creates a value of UpdateNotificationSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unsNotification - The notification specification for the HIT type.

unsActive - Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter for the call to UpdateNotificationSettings to succeed.

unsHITTypeId - The ID of the HIT type whose notification specification is being updated.

data UpdateNotificationSettings

See: updateNotificationSettings smart constructor.

Request Lenses

unsNotification :: Lens' UpdateNotificationSettings (Maybe NotificationSpecification)

The notification specification for the HIT type.

unsActive :: Lens' UpdateNotificationSettings (Maybe Bool)

Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter for the call to UpdateNotificationSettings to succeed.

unsHITTypeId :: Lens' UpdateNotificationSettings Text

The ID of the HIT type whose notification specification is being updated.

Destructuring the Response

updateNotificationSettingsResponse

Arguments

:: Int
unsrsResponseStatus

-> UpdateNotificationSettingsResponse

Creates a value of UpdateNotificationSettingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unsrsResponseStatus - -- | The response status code.

data UpdateNotificationSettingsResponse

See: updateNotificationSettingsResponse smart constructor.

Response Lenses

unsrsResponseStatus :: Lens' UpdateNotificationSettingsResponse Int

- | The response status code.