Network.AWS.MechanicalTurk.UpdateQualificationType

Description

The UpdateQualificationType operation modifies the attributes of an existing Qualification type, which is represented by a QualificationType data structure. Only the owner of a Qualification type can modify its attributes.

Most attributes of a Qualification type can be changed after the type has been created. However, the Name and Keywords fields cannot be modified. The RetryDelayInSeconds parameter can be modified or added to change the delay or to enable retries, but RetryDelayInSeconds cannot be used to disable retries.

You can use this operation to update the test for a Qualification type. The test is updated based on the values specified for the Test, TestDurationInSeconds and AnswerKey parameters. All three parameters specify the updated test. If you are updating the test for a type, you must specify the Test and TestDurationInSeconds parameters. The AnswerKey parameter is optional; omitting it specifies that the updated test does not have an answer key.

If you omit the Test parameter, the test for the Qualification type is unchanged. There is no way to remove a test from a Qualification type that has one. If the type already has a test, you cannot update it to be AutoGranted. If the Qualification type does not have a test and one is provided by an update, the type will henceforth have a test.

If you want to update the test duration or answer key for an existing test without changing the questions, you must specify a Test parameter with the original questions, along with the updated values.

If you provide an updated Test but no AnswerKey, the new test will not have an answer key. Requests for such Qualifications must be granted manually.

You can also update the AutoGranted and AutoGrantedValue attributes of the Qualification type.

Creating a Request

updateQualificationType

Arguments

:: Text
uqtQualificationTypeId

-> UpdateQualificationType

Creates a value of UpdateQualificationType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uqtTestDurationInSeconds - The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

uqtQualificationTypeStatus - The new status of the Qualification type - Active | Inactive

uqtAnswerKey - The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

uqtTest - The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified. Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true. Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

uqtAutoGranted - Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Constraints: If the Test parameter is specified, this parameter cannot be true.

uqtAutoGrantedValue - The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

uqtDescription - The new description of the Qualification type.

uqtRetryDelayInSeconds - The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

uqtQualificationTypeId - The ID of the Qualification type to update.

data UpdateQualificationType

See: updateQualificationType smart constructor.

Request Lenses

uqtTestDurationInSeconds :: Lens' UpdateQualificationType (Maybe Integer)

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

uqtQualificationTypeStatus :: Lens' UpdateQualificationType (Maybe QualificationTypeStatus)

The new status of the Qualification type - Active | Inactive

uqtAnswerKey :: Lens' UpdateQualificationType (Maybe Text)

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

uqtTest :: Lens' UpdateQualificationType (Maybe Text)

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified. Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true. Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

uqtAutoGranted :: Lens' UpdateQualificationType (Maybe Bool)

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Constraints: If the Test parameter is specified, this parameter cannot be true.

uqtAutoGrantedValue :: Lens' UpdateQualificationType (Maybe Int)

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

uqtDescription :: Lens' UpdateQualificationType (Maybe Text)

The new description of the Qualification type.

uqtRetryDelayInSeconds :: Lens' UpdateQualificationType (Maybe Integer)

The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

uqtQualificationTypeId :: Lens' UpdateQualificationType Text

The ID of the Qualification type to update.

Destructuring the Response

updateQualificationTypeResponse

Arguments

:: Int
uqtrsResponseStatus

-> UpdateQualificationTypeResponse

Creates a value of UpdateQualificationTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uqtrsQualificationType - Contains a QualificationType data structure.

uqtrsResponseStatus - -- | The response status code.

data UpdateQualificationTypeResponse

See: updateQualificationTypeResponse smart constructor.

Response Lenses

uqtrsQualificationType :: Lens' UpdateQualificationTypeResponse (Maybe QualificationType)

Contains a QualificationType data structure.

uqtrsResponseStatus :: Lens' UpdateQualificationTypeResponse Int

- | The response status code.